agents/part_definer.py
```python
from models.part import Part
from utils.openai_api import call_openai
from utils.google_search import search_google
from utils.webscraper import scrape_web
from utils.pinecone_vector_db import use_pinecone

class PartDefiner:
    def __init__(self, user_input):
        self.user_input = user_input
        self.part = Part()

    def define_part(self):
        self.part.subject = self.user_input.subject
        self.part.keywords = self.user_input.keywords
        self.part.target_content_type = self.user_input.target_content_type
        self.part.word_count = self.user_input.word_count

    def generate_content(self):
        # Use OpenAI API to generate content
        generated_content = call_openai(self.part.subject, self.part.keywords, self.part.target_content_type, self.part.word_count)
        self.part.content = generated_content

    def enrich_content(self):
        # Use Google Search and Web Scraper for non-fiction writing
        if self.part.target_content_type == 'non-fiction':
            search_results = search_google(self.part.keywords)
            scraped_data = scrape_web(search_results)
            self.part.content += scraped_data

    def embed_data(self):
        # Use Pinecone Vector Database for embedded data
        embedded_data = use_pinecone(self.part.content)
        self.part.embedded_data = embedded_data

    def get_part(self):
        return self.part
```